requirements.txt
asgiref==3.5.2
Django==4.0.6
django-admin-interface==0.19.1
django-colorfield==0.7.2
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
Pillow==9.2.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
